web/pyproject.toml
[project]
name = "web"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Ynakatsuka", email = "nk.tsssa@gmail.com" }
]
dependencies = [
    "streamlit>=1.28.1",
    "polars>=0.19.12",
    "scipy>=1.10.1",
    "plotly_express>=0.4.1",
    "yfinance>=0.2.31",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true
